+++ app/Http/Controllers/TaskStatusController.php
@@ -5,13 +5,10 @@ namespace App\Http\Controllers;
 use App\Models\TaskStatus;
 use App\Models\Task;
 use Illuminate\Http\Request;
-use Illuminate\Database\QueryException;
+use Illuminate\Database\Support\Facades\Auth;
 
 class TaskStatusController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
         $taskStatuses = TaskStatus::paginate();
@@ -19,18 +16,14 @@ class TaskStatusController extends Controller
         return view('task_statuses.index',  compact('taskStatuses'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
     public function create()
     {
-        $taskStatus = new TaskStatus();
-        return view('task_statuses.create', compact('taskStatus'));
+        if (Auth::guest()) {
+            return abort(403);
+        }
+        return view('task_statuses.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
         $data = $request->validate([
@@ -40,55 +33,37 @@ class TaskStatusController extends Controller
         $taskStatus = new TaskStatus();
         $taskStatus->fill($data);
         $taskStatus->save();
-        flash('Статус успешно создан')->success();
+        flash(__('controllers.task_statuses_create'))->success();
 
         return redirect()->route('task_statuses.index');
     }
 
-    /**
-     * Display the specified resource.
-     */
-    public function show(TaskStatus $taskStatus)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit($id)
+    public function edit(TaskStatus $taskStatus)
     {
         $taskStatus = TaskStatus::findOrFail($id);
         return view('task_statuses.edit', compact('taskStatus'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, $id)
+    public function update(Request $request, TaskStatus $taskStatus)
     {
-        $taskStatus = TaskStatus::findOrFail($id);
         $data = $request->validate([
             'name' => "required|unique:task_statuses,name,{$taskStatus->id}",
         ]);
         $taskStatus->fill($data);
         $taskStatus->save();
-        flash(__('Status successfully changed'))->success();
+        flash(__('controllers.task_statuses_update'))->success();
         return redirect()->route('task_statuses.index');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy($id)
+    public function destroy(TaskStatus $taskStatus)
     {
-        try {
-            $taskStatus = TaskStatus::find($id);
-            $taskStatus->delete();
-            flash('Статус успешно удален')->success();
-        } catch (QueryException $qe) {
-            flash('Не удалось удалить статус')->error();
+        if ($taskStatus->tasks()->exists()) {
+            flash(__('controllers.task_statuses_destroy_failed'))->error();
+            return back();
         }
+        $taskStatus->delete();
+
+        flash(__('controllers.task_statuses_destroy'))->success();
         return redirect()->route('task_statuses.index');
     }
 }
